+++ Runtime/Extensions/TransformExt.cs
@@ -1,0 +1,11 @@
+using ComradeVanti.CSharpTools;
+using UnityEngine;
+
+namespace Dev.ComradeVanti
+{
+    public static class TransformExt
+    {
+        public static IOpt<Transform> TryParent(this Transform transform) =>
+            transform.parent.ToOpt();
+    }
+}